+++ sentry-android-replay/src/test/java/io/sentry/android/replay/viewhierarchy/ComposeMaskingOptionsTest.kt
@@ -32,6 +32,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import coil.compose.AsyncImage
+import io.sentry.NoOpLogger
 import io.sentry.SentryOptions
 import io.sentry.android.replay.maskAllImages
 import io.sentry.android.replay.maskAllText
@@ -161,7 +162,12 @@ class ComposeMaskingOptionsTest {
       assertNotNull(composeView)
 
       val rootNode = GenericViewHierarchyNode(0f, 0f, 0, 0, 1.0f, -1, shouldMask = true)
-      ComposeViewHierarchyNode.fromView(composeView, rootNode, options)
+      ComposeViewHierarchyNode.fromView(
+        composeView,
+        rootNode,
+        options.sessionReplay,
+        NoOpLogger.getInstance(),
+      )
 
       assertEquals(1, rootNode.children?.size)
 
@@ -275,8 +281,8 @@ class ComposeMaskingOptionsTest {
 
   private inline fun <reified T> Activity.collectNodesOfType(options: SentryOptions): List<T> {
     val root = window.decorView
-    val viewHierarchy = ViewHierarchyNode.fromView(root, null, 0, options)
-    root.traverse(viewHierarchy, options)
+    val viewHierarchy = ViewHierarchyNode.fromView(root, null, 0, options.sessionReplay)
+    root.traverse(viewHierarchy, options.sessionReplay, NoOpLogger.getInstance())
 
     val nodes = mutableListOf<T>()
     viewHierarchy.traverse {