+++ sentry/src/main/java/io/sentry/SentryReplayOptions.java
@@ -11,25 +11,14 @@ import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Locale;
 import java.util.Set;
-import java.util.concurrent.CopyOnWriteArraySet;
 import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-public final class SentryReplayOptions {
+public final class SentryReplayOptions extends SentryMaskingOptions {
 
   private static final String CUSTOM_MASKING_INTEGRATION_NAME = "ReplayCustomMasking";
   private volatile boolean customMaskingTracked = false;
-
-  public static final String TEXT_VIEW_CLASS_NAME = "android.widget.TextView";
-  public static final String IMAGE_VIEW_CLASS_NAME = "android.widget.ImageView";
-  public static final String WEB_VIEW_CLASS_NAME = "android.webkit.WebView";
-  public static final String VIDEO_VIEW_CLASS_NAME = "android.widget.VideoView";
-  public static final String CAMERAX_PREVIEW_VIEW_CLASS_NAME = "androidx.camera.view.PreviewView";
-  public static final String ANDROIDX_MEDIA_VIEW_CLASS_NAME = "androidx.media3.ui.PlayerView";
-  public static final String EXOPLAYER_CLASS_NAME = "com.google.android.exoplayer2.ui.PlayerView";
-  public static final String EXOPLAYER_STYLED_CLASS_NAME =
-      "com.google.android.exoplayer2.ui.StyledPlayerView";
 
   /**
    * Maximum size in bytes for network request/response bodies to be captured in replays. Bodies
@@ -85,36 +74,6 @@ public final class SentryReplayOptions {
    * default is null (disabled).
    */
   private @Nullable Double onErrorSampleRate;
-
-  /**
-   * Mask all views with the specified class names. The class name is the fully qualified class name
-   * of the view, e.g. android.widget.TextView. The subclasses of the specified classes will be
-   * masked as well.
-   *
-   * <p>If you're using an obfuscation tool, make sure to add the respective proguard rules to keep
-   * the class names.
-   *
-   * <p>Default is empty.
-   */
-  private Set<String> maskViewClasses = new CopyOnWriteArraySet<>();
-
-  /**
-   * Ignore all views with the specified class names from masking. The class name is the fully
-   * qualified class name of the view, e.g. android.widget.TextView. The subclasses of the specified
-   * classes will be ignored as well.
-   *
-   * <p>If you're using an obfuscation tool, make sure to add the respective proguard rules to keep
-   * the class names.
-   *
-   * <p>Default is empty.
-   */
-  private Set<String> unmaskViewClasses = new CopyOnWriteArraySet<>();
-
-  /** The class name of the view container that masks all of its children. */
-  private @Nullable String maskViewContainerClass = null;
-
-  /** The class name of the view container that unmasks its direct children. */
-  private @Nullable String unmaskViewContainerClass = null;
 
   /**
    * Defines the quality of the session replay. The higher the quality, the more accurate the replay
@@ -276,60 +235,32 @@ public final class SentryReplayOptions {
     this.sessionSampleRate = sessionSampleRate;
   }
 
-  /**
-   * Mask all text content. Draws a rectangle of text bounds with text color on top. By default only
-   * views extending TextView are masked.
-   *
-   * <p>Default is enabled.
-   */
+  @Override
   public void setMaskAllText(final boolean maskAllText) {
-    if (maskAllText) {
-      maskViewClasses.add(TEXT_VIEW_CLASS_NAME);
-      unmaskViewClasses.remove(TEXT_VIEW_CLASS_NAME);
-    } else {
+    if (!maskAllText) {
       trackCustomMasking();
-      unmaskViewClasses.add(TEXT_VIEW_CLASS_NAME);
-      maskViewClasses.remove(TEXT_VIEW_CLASS_NAME);
     }
+    super.setMaskAllText(maskAllText);
   }
 
-  /**
-   * Mask all image content. Draws a rectangle of image bounds with image's dominant color on top.
-   * By default only views extending ImageView with BitmapDrawable or custom Drawable type are
-   * masked. ColorDrawable, InsetDrawable, VectorDrawable are all considered non-PII, as they come
-   * from the apk.
-   *
-   * <p>Default is enabled.
-   */
+  @Override
   public void setMaskAllImages(final boolean maskAllImages) {
-    if (maskAllImages) {
-      maskViewClasses.add(IMAGE_VIEW_CLASS_NAME);
-      unmaskViewClasses.remove(IMAGE_VIEW_CLASS_NAME);
-    } else {
+    if (!maskAllImages) {
       trackCustomMasking();
-      unmaskViewClasses.add(IMAGE_VIEW_CLASS_NAME);
-      maskViewClasses.remove(IMAGE_VIEW_CLASS_NAME);
     }
+    super.setMaskAllImages(maskAllImages);
   }
 
-  @NotNull
-  public Set<String> getMaskViewClasses() {
-    return this.maskViewClasses;
-  }
-
+  @Override
   public void addMaskViewClass(final @NotNull String className) {
     trackCustomMasking();
-    this.maskViewClasses.add(className);
+    super.addMaskViewClass(className);
   }
 
-  @NotNull
-  public Set<String> getUnmaskViewClasses() {
-    return this.unmaskViewClasses;
-  }
-
+  @Override
   public void addUnmaskViewClass(final @NotNull String className) {
     trackCustomMasking();
-    this.unmaskViewClasses.add(className);
+    super.addUnmaskViewClass(className);
   }
 
   @ApiStatus.Internal
@@ -361,28 +292,7 @@ public final class SentryReplayOptions {
     return sessionDuration;
   }
 
-  @ApiStatus.Internal
-  public void setMaskViewContainerClass(@NotNull String containerClass) {
-    maskViewClasses.add(containerClass);
-    maskViewContainerClass = containerClass;
-  }
-
-  @ApiStatus.Internal
-  public void setUnmaskViewContainerClass(@NotNull String containerClass) {
-    unmaskViewContainerClass = containerClass;
-  }
-
-  @ApiStatus.Internal
-  public @Nullable String getMaskViewContainerClass() {
-    return maskViewContainerClass;
-  }
-
-  @ApiStatus.Internal
-  public @Nullable String getUnmaskViewContainerClass() {
-    return unmaskViewContainerClass;
-  }
-
-  @ApiStatus.Internal
+  @Override
   public void trackCustomMasking() {
     if (!customMaskingTracked) {
       customMaskingTracked = true;